Move the functionalities from the OptimalSequence file in the WorkloadCDF
