+++ walltime_request/WorkloadCDF.py
@@ -245,6 +245,99 @@ class DistInterpolation(InterpolationModel):
 
         return (best_distribution, best_params, best_sse)
 
+#-------------
+# Classes for computing the sequence of requests
+#-------------
+
+class RequestSequence():
+    ''' Sequence that optimizes the total makespan of a job for discret
+    values (instead of a continuous space) '''
+
+    def __init__(self, max_value, discrete_values, probability_values):
+        self.discret_values = discrete_values
+        self.__prob = probability_values
+        self.upper_limit = max_value
+        self._E = {}
+        self._request_sequence = []
+        
+        self.__sumF = self.get_discret_sum_F()
+        E_val = self.compute_E_value(0)
+        self.__t1 = self.discret_values[E_val[1]]
+        self.__makespan = E_val[0]
+
+    def compute_F(self, vi):
+        fi = self.__prob[vi]
+        if vi > 0:
+            fi -= self.__prob[vi-1]
+        return fi / self.__prob[-1]
+
+    def get_discret_sum_F(self):
+        sumF = (len(self.discret_values) + 1) * [0]
+        for k in range(len(self.discret_values) - 1, -1, -1):
+            sumF[k] = self.compute_F(k) + sumF[k + 1]
+        return sumF
+
+    def __compute_E_table(self, i):
+        if i == len(self.discret_values):
+            return (0, len(self.discret_values))
+
+        min_makespan = -1
+        min_request = -1
+        for j in range(i, len(self.discret_values)):
+            makespan = float(self.__sumF[i] * self.discret_values[j])
+            if j + 1 in self._E:
+                makespan += self._E[j + 1][0]
+            else:
+                E_val = self.__compute_E_table(j + 1)
+                makespan += E_val[0]
+                self._E[j + 1] = E_val
+
+            if min_request == -1 or min_makespan > makespan:
+                min_makespan = makespan
+                min_request = j
+        return (min_makespan, min_request)
+
+    def __compute_E_table_iter(self, first):
+        self._E[len(self.discret_values)] = (0, len(self.discret_values))
+        for i in range(len(self.discret_values) - 1, first - 1, -1):
+            if i in self._E:
+                continue
+            min_makespan = 0
+            min_request = len(self.discret_values)
+            for j in range(i, len(self.discret_values)):
+                makespan = float(self.__sumF[i] * self.discret_values[j])
+                makespan += self._E[j + 1][0]
+
+                if min_makespan == 0 or min_makespan >= makespan:
+                    min_makespan = makespan
+                    min_request = j
+            self._E[i] = (min_makespan, min_request)
+        return self._E[first]
+
+    def compute_request_sequence(self):
+        if len(self._request_sequence) > 0:
+            return self._request_sequence
+        j = 0
+        E_val = self.compute_E_value(j)
+        while E_val[1] < len(self.discret_values):
+            self._request_sequence.append((self.discret_values[E_val[1]], ))
+            j = E_val[1] + 1
+            E_val = self.compute_E_value(j)
+
+        if self._request_sequence[-1][0] != self.upper_limit:
+            self._request_sequence.append((self.upper_limit, ))
+        
+        return self._request_sequence
+
+    def compute_E_value(self, i):
+        if i in self._E:
+            return self._E[i]
+        if len(self.discret_values)<600:
+            E_val = self.__compute_E_table(i)
+        else:
+            E_val = self.__compute_E_table_iter(i)
+        self._E[i] = E_val
+        return E_val
 
 #-------------
 # Classes for defining how the cost is computed
